refactor: Consolidate ones_like function and update test imports for consistency
